+++ docs/conf.py
@@ -18,8 +18,8 @@ else:
 project = "PDELie"
 author = "Alex Gabel"
 copyright = f"{date.today().year}, {author}"
-release = "0.28.0"
-version = "0.28"
+release = "0.29.0"
+version = "0.29"
 
 extensions = [
     "myst_parser",
@@ -48,6 +48,8 @@ html_theme_options = {
     "show_toc_level": 2,
     "navbar_align": "left",
 }
+
+latex_engine = "xelatex"
 
 myst_enable_extensions = [
     "colon_fence",